I think the model was overfitting to the training data, which is why the accuracy was low, so I added dropout layers to reduce overfitting.
No matter which values for the dropout I used, it made the accuracy worse instead of better.

I tried using 5x5 for the kernel_size, but that made the accuracy worse too.

I tried reducing the learning rate of RMSprop, but that lowered the accuracy too.

I tried raising the learning rate of RMSprop, but that severely lowered the accuracy.

I tried adding a decay rate to RMSprop, but no matter what value I chose, the accuracy stayed around 10% with the addition of a decay rate.

Found out the correct way to use a validation set by reading the Keras documentation. Made the necessary changes to do that.

A lot of work, a lot of changes, but the accuracy still hovers around 70%.